backend/requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt 
django-cors-headers 
django-filter 
drf-yasg 
weasyprint
whitenoise
gunicorn
psycopg2-binary
mysqlclient